infra/core/requirements.txt
# Sovereign Infrastructure - Season 2 Core Dependencies
# Gmail Local Agent

google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
